# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "مستندات API برای ثبت‌نام و احراز هویت کاربران",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/subscriptions": {
            "get": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "List user subscriptions (User)",
                "operationId": "24752636f8d944ceb934f9f9d92627be",
                "responses": {
                    "200": {
                        "description": "List of user subscriptions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscriptions retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubscriptionResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "Create a new subscription (User)",
                "operationId": "406c35c8f6e4b8a7856a1113b18cd233",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/{subscription}": {
            "get": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "Get subscription details (User)",
                "operationId": "54d2e5cb07cf221f420e794de286b07c",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "Update subscription (User)",
                "operationId": "aa157d7a274674617f852474194f8773",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "Cancel subscription (User)",
                "operationId": "85bd69947c9dda8634893190c65daa7e",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cancellation_reason": {
                                        "type": "string",
                                        "example": "Switching to a different plan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription cancelled successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/{subscription}/renew": {
            "post": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "Renew subscription (User)",
                "operationId": "02e681e1b275e1ef1ade815fa19b52c0",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription renewed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription renewed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/subscribe": {
            "post": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "خرید اشتراک جدید (User)",
                "operationId": "9db73788836f58abf906b24029c1d127",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "اشتراک با موفقیت ایجاد شد"
                                        },
                                        "subscription": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/subscriptions/{subscription}/cancel": {
            "post": {
                "tags": [
                    "Subscription (User)"
                ],
                "summary": "لغو اشتراک (User)",
                "operationId": "c2a109948b0f935e329b9c68d7e6d881",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cancellation_reason": {
                                        "type": "string",
                                        "example": "Switching to a different plan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subscription cancelled successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SubscriptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateSubscriptionRequest": {
                "title": "Update Subscription Request",
                "description": "درخواست بروزرسانی اشتراک",
                "properties": {
                    "auto_renew": {
                        "description": "آیا تمدید خودکار فعال باشد",
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "description": "وضعیت اشتراک",
                        "type": "string",
                        "enum": [
                            "active",
                            "cancelled"
                        ],
                        "example": "active"
                    },
                    "cancellation_reason": {
                        "description": "دلیل لغو اشتراک",
                        "type": "string",
                        "example": "درخواست کاربر",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "SubscriptionResource": {
                "title": "Subscription Resource",
                "description": "منبع اشتراک شامل اطلاعات کامل اشتراک",
                "properties": {
                    "id": {
                        "description": "شناسه اشتراک",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "شناسه کاربر",
                        "type": "integer",
                        "example": 1
                    },
                    "subscription_plan_id": {
                        "description": "شناسه طرح اشتراک",
                        "type": "integer",
                        "example": 1
                    },
                    "plan": {
                        "$ref": "#/components/schemas/SubscriptionPlanResource"
                    },
                    "start_date": {
                        "description": "تاریخ شروع اشتراک",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23T10:00:00Z"
                    },
                    "end_date": {
                        "description": "تاریخ پایان اشتراک",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-23T10:00:00Z"
                    },
                    "status": {
                        "description": "وضعیت اشتراک",
                        "type": "string",
                        "enum": [
                            "active",
                            "expired",
                            "cancelled",
                            "pending"
                        ],
                        "example": "active"
                    },
                    "payment_id": {
                        "description": "شناسه پرداخت",
                        "type": "string",
                        "example": "pay_123456789",
                        "nullable": true
                    },
                    "amount": {
                        "description": "مبلغ پرداختی",
                        "type": "number",
                        "format": "float",
                        "example": 100000
                    },
                    "payment_method": {
                        "description": "روش پرداخت",
                        "type": "string",
                        "example": "online",
                        "nullable": true
                    },
                    "payment_status": {
                        "description": "وضعیت پرداخت",
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "failed"
                        ],
                        "example": "completed"
                    },
                    "payment_date": {
                        "description": "تاریخ پرداخت",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23T10:00:00Z",
                        "nullable": true
                    },
                    "auto_renew": {
                        "description": "آیا تمدید خودکار فعال است",
                        "type": "boolean",
                        "example": true
                    },
                    "cancelled_at": {
                        "description": "تاریخ لغو اشتراک",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23T10:00:00Z",
                        "nullable": true
                    },
                    "cancellation_reason": {
                        "description": "دلیل لغو اشتراک",
                        "type": "string",
                        "example": "درخواست کاربر",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "تاریخ ایجاد اشتراک",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "تاریخ آخرین بروزرسانی اشتراک",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-23T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}